LayoutTests/fast/selectors/attribute-list-with-whitespace-in-value-expected.txt

When matching space separated tokens with the attribute selector [attr~=value], any of the following characters can be used as space separator: SPACE (U+0020), TAB (U+0009), LINE FEED (U+000A), CARRIAGE RETURN (U+000D), and FORM FEED (U+000C).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("[attr~=value]").length is document.querySelectorAll("target").length
PASS document.querySelectorAll("[attr~=value]").length is 21
PASS document.querySelectorAll("[attr~=value]")[0].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[1].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[2].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[3].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[4].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[5].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[6].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[7].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[8].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[9].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[10].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[11].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[12].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[13].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[14].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[15].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[16].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[17].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[18].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[19].tagName is "TARGET"
PASS document.querySelectorAll("[attr~=value]")[20].tagName is "TARGET"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[0]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[1]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[2]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[3]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[4]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[5]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[6]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[7]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[8]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[9]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[10]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[11]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[12]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[13]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[14]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[15]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[16]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[17]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[18]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[19]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[20]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[21]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[22]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[23]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[24]).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[25]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[26]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[27]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[28]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[29]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[30]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[31]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[32]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[33]).backgroundColor is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#testBlock *")[34]).backgroundColor is "rgb(0, 0, 0)"
PASS document.querySelectorAll("[attr~=\\002003 value]").length is 2
PASS document.querySelectorAll("[attr~=\\002003 value\\003000]").length is 1
PASS successfullyParsed is true

TEST COMPLETE
